--- src/components/page/StreamingPage/index.ts
@@ -1,1 +1,0 @@
-export { StreamingPage } from "./page";
