+++ src/service/streamer.ts
@@ -6,6 +6,7 @@ import { useClientInput } from "@/states/clientInput";
 export const useStreamService = () => {
     const { socket } = useSockets();
     const {
+        clientText,
         mutator: { updateText },
     } = useClientInput();
     const timeoutRef = useRef<number | null>(null);
@@ -26,7 +27,7 @@ export const useStreamService = () => {
     // 入力値が変更された時に実行される関数
     const handleInputChange = useCallback(
         (event: React.ChangeEvent<HTMLInputElement>) => {
-            const text = event.target.value;
+            const { value: text } = event.target;
             updateText(text);
             sendToServer(text);
         },
@@ -34,7 +35,8 @@ export const useStreamService = () => {
     );
 
     return {
-        timeoutRef,
+        socket,
+        clientText,
         handler: { handleInputChange },
     };
 };
